requirements.txt

google-cloud-aiplatform>=1.18.1
google-cloud-storage>=2.1.0
scikit-learn>=1.0.2
pandas>=1.3.5
numpy>=1.21.6
joblib>=1.1.0